Update style.css
I'd like you to please take note of the following information. TEXT: "Press spacebar to start the game" will not be visible in the output as the snake PNG overlaps the text. We need to move it down from the top to 65% to ensure visibility. The code to be used is:

#instruction-text{ 
position: absolute; 
top: 65%;
}